Update correct module loading according to restructured folders
